sam-sql-database-tool/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.sam_sql_database_tool.metadata]
type = "sam_sql_database_tool"

[project]
name = "sam_sql_database_tool"
version = "0.1.0"
description = "A reusable SQL database tool for Solace Agent Mesh."
readme = "README.md"
authors = [
    { name = "Hugo Pare", email = "hugo.pare@solace.com" },
]
requires-python = ">=3.10"
dependencies = [
    "SQLAlchemy==2.0.40",
    "PyMySQL==1.1.2",
    "psycopg2-binary==2.9.10",
    "pydantic==2.11.9",
    "PyYAML==6.0.2"
]

[tool.hatch.build.targets.wheel]
packages = ["sam_sql_database_tool"]
src-path = "src"

[tool.hatch.build.targets.wheel.force-include]
"src/sam_sql_database_tool" = "sam_sql_database_tool/"
"config.yaml" = "sam_sql_database_tool/config.yaml"
"README.md" = "sam_sql_database_tool/README.md"
"pyproject.toml" = "sam_sql_database_tool/pyproject.toml"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--tb=short --strict-markers --disable-warnings -p pytest_asyncio"

[tool.hatch.envs.hatch-test]
dependencies = [
  "a2a-sdk[http-server]>=0.3.1",
  "pytest-asyncio",
  "pytest>=8.0.0",
  "pytest-mock>=3.0.0",
  "pytest-cov>=4.0.0",
  "pytest-xdist>=3.5.0",
  "solace-agent-mesh>=1.4.9,<2.0.0",
  "testing.postgresql",
  "psycopg2-binary"
]
